Optimize run time cache initialization
We inline the run_time_cache in the map_ptr allocation instead of having inidividual pointers there.
This shall save a level of indirection, give better initialization properties, and possibly better data locality (all run_time_cache values from methods of a same class will now necessarily be together).

As to the initialization properties, the memory is now going to be cleared in a single fast operation: it is much more efficient to clear a larger chunk of memory at once than initializing many individual allocations.
This even holds true with opcache, where quite some code which is unused may be loaded, just comparing the speed of the two operations.