Allow using run_time_cache in Closures contained in classes with opcache
Without opcache, upon creation of the first closure, a shared run_time_cache entry was used.
With opcache however, given that ZEND_ACC_IMMUTABLE was always set and the scope was NULL by default, it did not happen with any closure contained in a class. This is significant, as the very most code nowadays resides in classes.
Also, given that non-static Closures generally inherit the same scope where they were defined in (with exception of traits), we can fixate the scope to the containing scope and have a matching scope on creation in the vast majority of the cases and use the run_time_cache of the closure.